added logic for modifying a trip

+++ src/api/repository/TripRepository.ts
@@ -1,5 +1,6 @@
 import { Types } from 'mongoose';
 import { Trip } from '../models/Trip';
+import { ModifyTripResponse } from './models/TripModels';
 import { TripModel } from './schemes/TripScheme';
 
 const createTrip = async (trip: Trip): Promise<Trip | null> => {
@@ -11,21 +12,74 @@ const getTrip = async (tripId: string): Promise<Trip | null> => {
   return await TripModel.findById(tripId);
 };
 
-//TODO add search for tickers, titles, or descriptions.
 const getTrips = async (): Promise<Trip[]> => {
   return await TripModel.find();
 };
 
-const upadateTrip = async (tripId: string): Promise<boolean> => {
-  const res = await TripModel.findOneAndUpdate({ tripId });
-  //retruns boolean for if it is successfull or not
-  return res !== null;
+const upadateTrip = async (
+  tripId: string,
+  managerId: string,
+  trip: Trip
+): Promise<ModifyTripResponse> => {
+  const doc = await TripModel.findOne({ _id: tripId });
+
+  const response: ModifyTripResponse | null = isTripModifiable(
+    tripId,
+    managerId,
+    doc
+  );
+
+  if (response) {
+    return response;
+  }
+
+  if (!doc) {
+    return {
+      statusCode: 404,
+      message: `Could not find Trip with Id: ${tripId}`,
+      isModified: false,
+    };
+  }
+
+  doc.overwrite(trip);
+  await doc.save();
+  return {
+    isModified: true,
+    statusCode: 200,
+    message: 'Trip is updated',
+  };
 };
 
-const deleteTrip = async (tripId: string): Promise<boolean> => {
+const deleteTrip = async (
+  tripId: string,
+  managerId: string
+): Promise<ModifyTripResponse> => {
+  const trip: Trip | null = await TripModel.findById(tripId);
+
+  const response: ModifyTripResponse | null = isTripModifiable(
+    tripId,
+    managerId,
+    trip
+  );
+
+  if (response) {
+    return response;
+  }
+
   const res = await TripModel.deleteOne({ _id: tripId });
-  //retruns boolean for if it is successfull or not
-  return res.deletedCount > 0;
+  if (res.deletedCount < 0) {
+    return {
+      isModified: false,
+      message: 'Could not modify trip',
+      statusCode: 500,
+    };
+  }
+
+  return {
+    isModified: true,
+    message: 'Trip deleted',
+    statusCode: 200,
+  };
 };
 
 const getTripsByActor = async (actorId: string): Promise<Trip[] | null> => {
@@ -33,19 +87,52 @@ const getTripsByActor = async (actorId: string): Promise<Trip[] | null> => {
     return null;
   }
 
-  console.log(actorId);
+  const trips = await TripModel.find({
+    'applications.actorId': actorId,
+  });
 
-  const applicationSearch = [
-    {
-      $unwind: '$applications',
-    },
-    {
-      $match: {
-        'applications.actorId': new Types.ObjectId(actorId),
-      },
-    },
-  ];
-  const trips = await TripModel.find(applicationSearch);
+  return trips;
+};
+
+const isTripModifiable = (
+  tripId: string,
+  managerId: string,
+  trip: Trip | null
+): ModifyTripResponse | null => {
+  if (trip === null) {
+    return {
+      isModified: false,
+      message: `Trip with Id: ${tripId} does not exist`,
+      statusCode: 404,
+    };
+  }
+
+  if (trip.managerId !== managerId) {
+    return {
+      isModified: false,
+      message: 'You have to be the manager for this trip to modify it',
+      statusCode: 403,
+    };
+  }
+
+  if (trip.isPublished) {
+    return {
+      isModified: false,
+      message: 'You cannot modify a trip that has been published',
+      statusCode: 405,
+    };
+  }
+  return null;
+};
+
+const getTripsByManager = async (managerId: string): Promise<Trip[] | null> => {
+  if (!Types.ObjectId.isValid(managerId)) {
+    return null;
+  }
+
+  const trips = await TripModel.find({
+    managerId: managerId,
+  });
 
   return trips;
 };
@@ -57,4 +144,5 @@ export const TripRepository = {
   getTrips,
   upadateTrip,
   getTripsByActor,
+  getTripsByManager,
 };
